Split initial margin requirements into buy and sell, use for OG risk checks (#661)
This change allows the order gateway to compare the margin with and
without an order to see if the new order would exceed the initial margin
requirement.

- (feat) OG now computes initial margin with/without an order as part of
risk checks

+++ src/protocol/api_gateway.rs
@@ -424,7 +424,7 @@ pub struct Balance {
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
 #[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
 pub struct SymbolRiskSnapshot {
-    pub open_quantity: Decimal,
+    pub open_quantity: i64,
     pub open_notional: Decimal,
     pub average_price: Option<Decimal>,
     pub initial_margin_required_position: Decimal,
